Fix namespace mess and server start up.
Fix undefiend method usage for starting the server.
Pass needed host argument to method that starts the server.
Big rename of the namespace variables for each module. According to
http://nodejs.org/docs/v0.6.1/api/modules.html each variable is only visible
inside a module unless it's added to "exports". So we don't need to define the
webinos.session.* namespaces in each module again and again.