Groundwork for instantiating services on pzp.
Spit out the browser code for the ServiceDiscovery into its own file
as we can then require this file in nodejs which runs the pzp.
Also the ServiceDiscovery is now a instantiable like all webinos
services. Which means it doesn't use the a static rpc handler
interface now, but a instance variable for the rpc handler.