Fixed the bug that causes event API not working properly.
It was a general bug, node returns multiple packets in one data event on client side (PZP). So we have to go through messages to find if it contains multiple packets and process them one by one.

Also removed process.nexttick as node code just skips insideticket if socket is in pause state