Build multiple impl apk versions according to platform version
Jira issue: WP-891

This contains a significant restructuring of the relationship between
the core webinos runtime and the api implementations in the impl
project. The app project now omits the the api and impl projects,
and those are now separately built into an apk whose contents depends
on the target platform version.